+++ packages/daucus/cli/index.js
@@ -1,2 +1,7 @@
-// TODO: CJS ?
+/**
+ * @typedef {import('./src/config/DaucusConfig').DaucusJSConfig} DaucusJSConfig
+ * @typedef {import('./src/commands/build.options').BuildCommandOptions} BuildCommandOptions
+ */
+
 export { BuildCommand } from "./src/commands/build.js";
+export { WorkSpace } from "./src/config/workspace.js";