+++ packages/daucus/cli/src/commands/build.options.d.ts
@@ -1,0 +1,10 @@
+/**
+ * parameters for the "build" command
+ *
+ */
+export interface BuildCommandOptions {
+  /**
+   * Name of the project to build (by default, all projects will be built)
+   */
+  project?: string;
+}