+++ packages/daucus/pandoc/index.doc.d.ts
@@ -1,0 +1,4 @@
+export { PandocOutputFormat, PandocInputFormat } from "./src/io-formats";
+
+export { run as pandoc } from "./src/run.js";
+export { md2html, convert } from "./src/convert.js";