+++ packages/daucus/pandoc/index.js
@@ -1,2 +1,7 @@
+/**
+ * @typedef {import("./src/io-formats").PandocInputFormat} PandocInputFormat
+ * @typedef {import("./src/io-formats").PandocOutputFormat} PandocOutputFormat
+ */
+
 export { run as pandoc } from "./src/run.js";
 export { md2html, convert } from "./src/convert.js";