+++ website/daucus.config.mjs
@@ -1,5 +1,9 @@
-export default {
+/** @type {import('@daucus/cli').DaucusJSConfig}*/
+const config = {
   defaultCompiler: "pandoc",
+  defaultCompilerOptions: {
+    filters: ["pandoc-import-code"],
+  },
   output: "src/fragments/",
   i18n: true,
   projects: {
@@ -11,5 +15,24 @@ export default {
       src: "**/*.md",
       root: "blog",
     },
+    daucus: {
+      src: "**/*.md",
+      root: "../packages/daucus",
+      exclude: [
+        "**/node_modules/**",
+        "**/fixtures/**",
+        "**/snowpack-template/**",
+      ],
+    },
+    helpers: {
+      src: "**/*.md",
+      root: "../packages/helpers",
+    },
+    "custom-element-name": {
+      src: "**/*.md",
+      root: "../packages/custom-element-name",
+    },
   },
 };
+
+export default config;