+++ packages/daucus/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.2.0](https://github.com/fullwebdev/fullwebdev/compare/daucus@v0.1.0...daucus@v0.2.0) (2021-04-25)
+
+### Features
+
+- **daucus:** typedoc-theme ([13bcbf4](https://github.com/fullwebdev/fullwebdev/commit/13bcbf48eca20c79e8dd8bbae782589dc0c0f2c8)), closes [#75](https://github.com/fullwebdev/fullwebdev/issues/75)
+- **daucus-cli:** compiler options ([a653ba2](https://github.com/fullwebdev/fullwebdev/commit/a653ba254426a61d40c4a70ac77a778e9c965d00)), closes [#74](https://github.com/fullwebdev/fullwebdev/issues/74)
+- **daucus-cli:** exclude project source files ([3e15faf](https://github.com/fullwebdev/fullwebdev/commit/3e15fafad9af74f2163b421d61d0040a62495a0b))
+- **daucus-cli:** help command ([17a8c3a](https://github.com/fullwebdev/fullwebdev/commit/17a8c3a00c3c7c36b316199bd7f680e0ecc67568))
+- **daucus-cli:** json configuration schema ([fcbea01](https://github.com/fullwebdev/fullwebdev/commit/fcbea019108c23a80f4eb03c4fa87e2b19690f60))
+- **daucus-cli:** more permissive function compilers ([2343fd2](https://github.com/fullwebdev/fullwebdev/commit/2343fd23ee32ab90e0574f0d5117892353faf1fb))
+- **daucus-cli:** rewrite relative urls ([0213bb3](https://github.com/fullwebdev/fullwebdev/commit/0213bb329301e85c228fa2309ecd6aff616c95d7))
+- **daucus-cli:** use filepath for route title ([3576979](https://github.com/fullwebdev/fullwebdev/commit/35769795e983ac47ce3e439d49ac29c9cb41a368))
+- **daucus-router:** URL fragment & params ([6b4f494](https://github.com/fullwebdev/fullwebdev/commit/6b4f4948471c0ddeaadea471c3f84291a4aaedd5)), closes [#76](https://github.com/fullwebdev/fullwebdev/issues/76)
+
+### Bug Fixes
+
+- **daucus-cli:** configuration typings ([8d81cf0](https://github.com/fullwebdev/fullwebdev/commit/8d81cf0fb6cf4b008d0c4e8057055708cd97dcf2))
+- **daucus-cli:** sort menu entries without prefix ([d5f270a](https://github.com/fullwebdev/fullwebdev/commit/d5f270ad557af15f758277b3e610e1c673d657a9))
+- **daucus-pandoc:** remove dependency to Python ([70857ce](https://github.com/fullwebdev/fullwebdev/commit/70857ce9e52dfbfdc69b56ed9272602a558e9fd2)), closes [#74](https://github.com/fullwebdev/fullwebdev/issues/74)
+- **daucus-router:** dependency to @modern-helpers/router 0.3 ([701bb8f](https://github.com/fullwebdev/fullwebdev/commit/701bb8fd2ae37f2d8538e1d50d33d094a8fa9d4e))
+- **daucus-router:** typings and custom-elements.json ([cf4c484](https://github.com/fullwebdev/fullwebdev/commit/cf4c4840951743be3273ededc3ff471260065430))
+- **daucus-router:** use @modern-helpers/router 0.4 ([a90d729](https://github.com/fullwebdev/fullwebdev/commit/a90d72962dcb510e1b9bce5ddf93c2224837c8c5))
+- **daucus-snowpack-template:** daucus config typings ([ec88397](https://github.com/fullwebdev/fullwebdev/commit/ec88397d5c4501ebda1c61d17c9426c5c4137422))
+- **html-loader:** update custom-elements.json ([d694518](https://github.com/fullwebdev/fullwebdev/commit/d694518cb91e7933798ae3a68837514505cb0834))
+
 ## 0.1.0 (2021-02-19)
 
 ### Features